refactoring based on internal discussion (#40)
